publisher_service/requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
Deprecated==1.2.15
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-pubsub==2.27.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
idna==3.10
importlib_metadata==8.5.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
